Acutally, just consolidate the api call classes within their modules, the tests make more sense this way.
